backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.7
python-decouple==3.8
Pillow==10.0.1
reportlab==4.0.4
django-filter==23.3
requests==2.31.0
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
sib-api-v3-sdk==7.6.0
httpx==0.25.2
django-redis==5.4.0
redis==5.0.1